Added code for characterSelection and Optimized Game Structure
The character selection code was created in case we decide to add customization to the game. For the game optimization, we saved the data of the repository, token and the game structure in DataManager. Based on the tests done the game works. Will later compare the times of the game from previous version with the new version.